feat(types): consolidate space sharing types into a single file and update imports
- Moved space sharing related types to a new file for better organization.
- Updated imports across the codebase to reference the new type definitions.
- Refactored API calls in the invite page to utilize the new unified API functions.
