+++ src/types/index.ts
@@ -1,0 +1,14 @@
+/**
+ * 類型定義匯出
+ * 統一匯出常用的類型定義
+ */
+
+// Prompt 相關類型
+export * from './prompt';
+export * from './fieldGropProps';
+export * from './nav';
+
+// 共用類型
+export * from '../shared/types/promptSpace';
+export * from '../shared/types/spaceSharing';
+export * from '../shared/types/prompt';
